Extract CredentialEncrypter to separate module
