+++ auth/auth-util/src/main/java/com/bazaarvoice/emodb/auth/util/CredentialEncrypter.java
