+++ blob-clients-common/blob-jersey2-client-common/pom.xml
@@ -47,6 +47,11 @@
             <artifactId>emodb-common-jersey2-client</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>com.bazaarvoice.emodb</groupId>
+            <artifactId>emodb-auth-util</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- Third party dependencies -->
         <dependency>
@@ -72,11 +77,6 @@
         <dependency>
             <groupId>dev.failsafe</groupId>
             <artifactId>failsafe</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.bazaarvoice.emodb</groupId>
-            <artifactId>emodb-auth-core</artifactId>
-            <version>${project.version}</version>
         </dependency>
     </dependencies>
 </project>
