+++ sor-client2/src/main/java/com/bazaarvoice/emodb/sor/client/DataStoreAuthenticator.java
@@ -2,7 +2,7 @@ package com.bazaarvoice.emodb.sor.client;
 
 import com.bazaarvoice.emodb.auth.InvalidCredentialException;
 import com.bazaarvoice.emodb.auth.proxy.CachingAuthenticatingProxy;
-import com.bazaarvoice.emodb.auth.util.ApiKeyEncryption;
+import com.bazaarvoice.emodb.auth.util.CredentialEncrypter;
 import com.bazaarvoice.emodb.sor.api.AuthDataStore;
 import com.bazaarvoice.emodb.sor.api.DataStore;
 
@@ -23,7 +23,7 @@ public class DataStoreAuthenticator extends CachingAuthenticatingProxy<DataStore
         if (apiKey.isEmpty()) {
             throw new InvalidCredentialException("API key cannot be empty");
         }
-        if(ApiKeyEncryption.isPotentiallyEncryptedApiKey(apiKey)){
+        if(CredentialEncrypter.isPotentiallyEncryptedString(apiKey)){
             throw new InvalidCredentialException("API Key is encrypted, please decrypt it");
         }
         return apiKey;
